get rid of keybase.socket and kbfs.socket
Having systemd open sockets for us is cool, but it was causing
unfortunate corner cases when the user switched back and forth between
KEYBASE_SYSTEMD=1 and KEYBASE_SYSTEMD=0 within the same session. This
mostly impacts just me, I think, cause I'm always toggling this setting,
but it hits me often enough that I don't want to risk it in the wild.